+++ dev/Examples.vue
@@ -38,6 +38,7 @@ import Job from '../components/Job.vue';
 import JobEstimate from '../components/JobEstimate.vue';
 import JsonSchema from '../components/JsonSchema.vue';
 import LinkList from '../components/LinkList.vue';
+import Logs from '../components/Logs.vue';
 import ObjectTree from '../components/ObjectTree.vue';
 import Process from '../components/Process.vue';
 import Processes from '../components/Processes.vue';
@@ -70,6 +71,7 @@ export default {
 		JobEstimate,
 		JsonSchema,
 		LinkList,
+		Logs,
 		ObjectTree,
 		Process,
 		Processes,