+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2020-05-05
+
 ### Added
 
 - New `Job` component
@@ -230,7 +232,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.3...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.3...v2.1.0
 [2.0.3]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.0.0...v2.0.1